requirements.txt
Authlib==1.7.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.7
dnspython==2.8.0
Flask==3.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
gunicorn==25.3.0
idna==3.12
itsdangerous==2.2.0
Jinja2==3.1.6
joserfc==1.6.4
MarkupSafe==3.0.3
packaging==26.1
pycparser==3.0
pymongo==4.17.0
python-dotenv==1.2.2
requests==2.33.1
SQLAlchemy==2.0.49
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
